chore: rename auth forder name to security
